track on table personel fix

+++ app/Http/Controllers/Admin/PersonelCrudController.php
@@ -757,6 +757,8 @@ class PersonelCrudController extends CrudController
             $request->request->set('valid_card_end',null);
         }
 
+        $errors = [];
+
         $this->crud->setRequest($request);
         $this->crud->unsetValidation(); // Validation has already been run
 
@@ -813,6 +815,68 @@ class PersonelCrudController extends CrudController
                 $isDuplicate->where('date_of_birth', $request->date_of_birth);
             }
 
+            $church_name = false;
+            $input_church_name = false;
+
+
+            if(preg_match_all('/(\{[\:\"\_\,a-z0-9]+\})/', $model->church_name, $matches)) {
+                $church_name = $matches[1];
+            }
+
+            if(preg_match_all('/(\{[\:\"\_\,a-z0-9]+\})/', $request->input("church_name"), $matches)) {
+                $input_church_name = $matches[1];
+            }
+
+            $trigger_matches_church = 0;
+            if(is_array($input_church_name)){
+                foreach($input_church_name as $church){
+                    if(!in_array($church, $church_name)){
+                        $trigger_matches_church = 1;
+                    }
+                }
+            }
+
+            $count_church_name_data = array_count_values($input_church_name);
+
+
+            // check validation double data church_name
+
+            
+            if($count_church_name_data){
+                foreach($count_church_name_data as $key => $number){
+                    if($number > 1){
+                        $errors['church_name'] = ['The pastor with same church Name has already exists.'];
+                    }
+                }
+            }
+
+
+            
+
+            $isDuplicate = $isDuplicate->select('id')->first();
+
+            if ($isDuplicate != null && $isDuplicate->id != $id) {
+                // DB::rollback();
+                // $errors = [
+                //     'first_name' => ['The pastor with same First Name, Last Name, Church Name and Date of Birth has already exists.'],
+                //     'last_name' => ['The pastor with same First Name, Last Name, Church Name and Date of Birth has already exists.'],
+                //     // 'church_name' => ['The pastor with same First Name, Last Name, Church Name and Date of Birth has already exists.'],
+                //     'date_of_birth' => ['The pastor with same First Name, Last Name, Church Name and Date of Birth has already exists.'],
+                // ];
+                $errors['first_name'] = ['The pastor with same First Name, Last Name, Church Name and Date of Birth has already exists.'];
+                $errors['last_name'] = ['The pastor with same First Name, Last Name, Church Name and Date of Birth has already exists.'];
+                $errors['date_of_birth'] = ['The pastor with same First Name, Last Name, Church Name and Date of Birth has already exists.'];
+
+                // return redirect($this->crud->route . '/'. $id . '/edit')
+                //     ->withInput()->withErrors($errors);
+            }
+
+            if(count($errors) > 0){
+                DB::rollback();
+                return redirect($this->crud->route . '/'. $id . '/edit')
+                ->withInput()->withErrors($errors);
+            }
+
             if ($model->church_name != $request->input("church_name")) {
                 StructureChurch::where('personel_id', $model->id)->delete();
                 $leaderships = json_decode($request->input("church_name"));
@@ -826,37 +890,15 @@ class PersonelCrudController extends CrudController
                     }
                 }
             }
-            
 
-            $isDuplicate = $isDuplicate->select('id')->first();
 
-            if ($isDuplicate != null && $isDuplicate->id != $id) {
-                DB::rollback();
-                $errors = [
-                    'first_name' => ['The pastor with same First Name, Last Name, Church Name and Date of Birth has already exists.'],
-                    'last_name' => ['The pastor with same First Name, Last Name, Church Name and Date of Birth has already exists.'],
-                    // 'church_name' => ['The pastor with same First Name, Last Name, Church Name and Date of Birth has already exists.'],
-                    'date_of_birth' => ['The pastor with same First Name, Last Name, Church Name and Date of Birth has already exists.'],
-                ];
-                return redirect($this->crud->route . '/'. $id . '/edit')
-                    ->withInput()->withErrors($errors);
+            // hit api for update user
+            if($com || ($trigger_matches_church == 1)){
+                $send = new HitApi;
+                $id = [$com];
+                $module = 'user_admin';
+                $response = $send->action($id, 'update', $module)->json();
             }
-            // if ($request->input('is_lifetime') == 0 && $request->input('valid_card_end') == null) {
-            //     DB::rollback();
-            //     $errors = [
-            //         'valid_card_end' => ['Valid Card End is required while lifetime is unchecked']
-            //     ];
-            //     return redirect($this->crud->route . '/'. $id . '/edit')
-            //         ->withInput()->withErrors($errors);
-            // }
-
-
-            // if($com){
-            //     $send = new HitApi;
-            //     $id = [$com];
-            //     $module = 'user_admin';
-            //     $response = $send->action($id, 'update', $module)->json();
-            // }
             
             // update the row in the db
             $item = $this->crud->update($request->get($this->crud->model->getKeyName()),
@@ -907,14 +949,7 @@ class PersonelCrudController extends CrudController
             // }
 
             DB::commit();
-             // hit api for update user
-             if($com){
-                $send = new HitApi;
-                $id = [$com];
-                $module = 'user_admin';
-                $response = $send->action($id, 'update', $module)->json();
-             }
-        } catch (Exception $e) {
+        } catch (Exception $e) { 
             DB::rollback();
             throw $e;
         }
